+++ CHANGELOG.md
@@ -3,6 +3,7 @@ medullah-web changelog file
 
 ## 0.20.7 (2024-09-25)
 * feat(rabbitmq): "close" method to close connection
+* feat(rabbitmq): acquire connection pool in use by this instance
 
 ## 0.20.6 (2024-09-25)
 * bump(crates): to their respective latest versions